Add binary to print parse tokens
For quickly running on actual source files, an executable that reads
files and prints out list of tokens is generated in Main.hs and related
files are modified to move some of the testing behaviour inside the
FixedForm lexer as it would be the same code to get a list of tokens.

In the future this binary will be extended to free form lexer as well as
various parsers and at some point will start using a pretty printer for
easy inspection.