Add action as Maybe Token and polymorphic lexer
Making actions Maybe Token allows FixedLex monads to return Nothing if
they do not intend to return Token such as those that change start
codes. The lexer itself upon seeing Nothing then calls the lexer' again
or the lexer itself may return Nothing (enclosed with the FixedLex
monad), which indicates failure, but makes it easy to test as the we can
generate test cases with quick check and say "in this context either we
return a label or and error." Previously this was not possible as
catching the exception results in being trapped in side the IO monad.

The lexer needs to be polymorphy with signature (Token -> Parse
AlexInput a) -> a. This was not possible as alexScanUser returns
a bounded FixedLex and when we use that in the existing monad, we would
get an error. That prevents signature to be polymorphic on a.
unsafeCoerce# is used to coerce the action into polymorphic type. This
makes sense as all possible values of action as defined by us are
polymorphic in a.