Add more states to lexer, support ENTRY
Add support for ENTRY statement.

Lexer has different modes now, which allows it to lex keywords as
identifiers when they cannot be keywords in that context such as in an
expression.

A "keyword" state is added for the first token of every statement. There
are two possibilities. Either it will start with a keyword proper, or it
will be an expression assignment. To differentiate keyword assignment do
an infinite look ahead within the statement to find a "=" token. If it
is find assume the token is an identifier, if not it is a keyword.

The problem this causes is with the DO and logical IF statements. DO
contains a statement itself with "=" token. IF may contain it. This problem
(for now) is resolved by not allowing DO or IF to be used as variable names
when they are at the beginning of the statement. This needs to be fixed
later on.

Additionally, iif state is added to go back to keyword state once the
parantheses after IF keyword is balanced. This parantheses count is
added to the parse state.