Restructure disambiguator to transformer
There will be multiple transformations for the purposes of structuring
the AST as in the case of IF-THEN-ELSE block structuring, as well as
disambiguation using type/scoping information. Hence, a generic
transformer module allows parsers or other AST users to choose which
transformations to apply.

While at it added some hspec tests for grouping.