Fix equals look ahead bug
Do differentiate between a keyword and an ID, the lexer did a lookahead
for an assignment (=), but it was too loose that things like reads,
writes, and logical if statements gave false positives.

The bug is fixed by either looking for a "=" immediately after the
supposed ID token, or well paranthesised list (representing array
element access) immediately followed by a "=".