Top of AST goes from [ProgramUnit] to ProgramFile
To express interleaving comments between program units in the AST, the
top of the AST is changed to ProgramFile, which holds pairs of comment
program units and an optional epilogue of comments.

Additionally there was a bug regarding seeing a function with its type
given, as after lexing the type the lexer goes back to statement state,
that is changed to lookahead for a Function and prevent start code
change in that case.

While doing that realised under lookahead cases lexer function is used,
hence if something goes wrong due to our assumption. We do not backtrack
and return false instead, we exit. All occurances of such are replaced
with lexerM, which return a maybe monad in case of failure. So we can
backtrack properly.