Derive Functor for all AST data types
This implies that (AList t a) must only be used when t is a functor (and
therefore it is a type constructor of kind * -> * instead of the previous case
where it was of kind *). The code has been modified so that AList is only used
on functors. The main victim of this change has been arguments to functions,
which were previously represented as Strings, but Strings aren't functors. So
now they are represented as the ValVariable constructor in the Value datatype.