Fix arithmetic if lexing bug.
Parantheses count was used to be able to lex statements after the if
condition. This was buggy in free form lexer as matched parantheses
after the if condition also causes said start code transition. To
prevent it a boolean flag is added to keep track of if the change has
occured. If it did, the following transitions 0 crossings don't cause a
start code change.