Fix function disambiguation
Function disambiguation was failing due to a false assumption, which was
the every global namespace will have an entry in the outer type mapping,
this is not true if there are no information sources for such function
calls/arrays. This was fixed by making result of each lookup an instance
of a Maybe monad and ignoring the mapping if none exists!

Meanwhile the whole transformation was transformed from monomorphic to
polymorphic.

Note that Global namespace should still need definition of intrinsic
functions for proper disambiguation.