Fix comment lexing bug
The bug was due to the fact that single ! pattern is too short so other
patterns are matched instead of it. When the pattern is changed to !.*\n
the problem is with & characters inside the comment as they are
interpreted as continuation starters. To remedy that the condition for
continuation is ammended to check if the lexeme starts with !, which is
possible only if the token being parsed is a comment.