Remove annotation from ValVariable. Add: varName, genVar functions.
varName should be used instead of pattern-matching to extract the name of an
ExpValue variable. This is because it will first return the uniqueName if the
variable has one. This is better for doing analysis; otherwise you must rewrite
all of the ValVariables with the uniqueName before applying any analyses.

genVar is a helper function to generate a new ExpValue variable with
annotation, source span and uniqueName == source name.