Hack to prevent arrays to be disambiguated.
Since in Fortran 90, the original function disambiguation routine is not
sound. Arrays may be inccorrectly deduced as functions. As a result
strides are attempted to be converted to arguments, which result in
failures. This adds a separate clause that ensures if indicies of a
deduced function has a stride, it remains as an array.

This is temporary until a sound replacement is written.