Enhance Block AST nodes with reconstruction info
Grouped blocks were lacking the label for the final statement, which is
something like `end do`, `end select`, etc. This was compansated by
adding the last block directly to the lasy body of the grouped block.
However, during printing we don't want to depend on this hack, so
instead the label is stored in grouped block nodes, to be recreated
later.

Also construct name was missing from DO, WHILE, and IF blocks. That is
also fixed.