genBackEdgeMap: Convert to BackEdgeMap only after filtering for dominance.
Conversion to BackEdgeMap too soon screws up one-to-many nodes, since it is a
one-to-one map. This wasn't a problem in the past because of the way we
translated loops into AST, the 'continue' basic block was always a 1-entrance,
1-exit deal. But now it is possible for a do-loop to be translated into a
single basic block with multiple exits, so the bug is triggered.