v0.2.3.0: refactor functions/subroutines, properly support prefixes and suffixes.
PUFunctionOpts has been replaced with PrefixSuffix, which is two ALists:
Prefixes are things like ELEMENTAL, RECURSIVE and PURE.
Suffixes are, for now, BIND (...) in Fortran2003 only.

Note that strictly speaking the return type of the function should be a prefix
and the result variable (if specified) should be a suffix, however I have left
these in their current arrangement so as not to disturb the AST. They are
special-cased.