Promote FortranVersion to own separate module
`FortranVersion` is used by various parts of the project and by
fortran-src users, so placed into its own top-level module.

For now, `ParserMonad` continues to export it in order to keep API from
breaking. That is, `import Language.Fortran.ParserMonad (
FortranVersion(..) )` will continue to work, but it should be consider
deprecated.