Fix assignment to array in a struct
This is a redo of a patch by @aburzillo to correctly analyse the type
of data references, due to the changes to fortran-src types and the new
`SemanticType`. Given the plan to eventually use the information from
this stage in the type checker this partial support for structures will
need to be extended.