fix some pretty print errors
Structures are only parsed in the Fortran77Legacy parser. However, their
pretty printing requests that they are Fortran77Extended, or it fails
with a `tooOld Fortran77Extended` call. This doesn't make sense
(77Legacy is "newer" than 77Extended), so I replaced it with `olderThan
Fortran77Legacy` -- perhaps not ideal, but I didn't want to start a
pattern of specifying a list of versions to compare against.

Also, common groups now have better spacing.