parse REAL literals to convenient form
We decompose REALs to their respective parts (significand & exponent),
but keep the numeric parts as strings to allow the user to select how to
represent the value. Along with a printer and parser, we also export a
"reader" to convert parsed REAL literals to any compatible Fortran type,
so for most uses users will only need that function to turn a Fortran
REAL literal into a Haskell value. (Previously, everyone had to
implement their own parsing.)