attempt to parse bracketed plain vars specially
Based on the following tweet and referenced material

    call double((i), i)

has potentially different meaning to, and should be parsed differently
from

    call double(i, i)

As I understand, variables may be used as expressions in two ways:
either directly, as a value access, or indirectly, as an expression to
be evaluated. I add a new Expression constructor for the latter, throw
some hopefully-overriding parser rules in and hope for the best.

https://twitter.com/fortrantip/status/1479071485859962880