add named constants to constant literal part
Explicitly added in F2003 standard. For now, we use the same parsing for
all versions - it seems to be one of those features that is silently
accepted by most compilers...

Also refactored some AST definitions: more definitions are placed in
their own module, and the literals have been grouped together. The AST
module is incredibly slow to compile, I imagine due to the volume of
instances generated. This should help.