use generic-optics to define First,SecondParameter
Our generics code was unsafe in that if you tried to derive an instance
which was invalid (e.g. you say an `a` is the first parameter in each
constructor, but it isn't), it would still derive an instance, but give
you an `undefined` runtime error upon using it. Pushing that error to
compile time is complex, but it's a solved problem by kcsongor's in his
wonderful generic-lens library.

This change adds generic-optics and optics-core to dependencies. They're
very lightweight. Performance should be the same; compile time should
most likely increase.