SemType: replace Dimension to support assumed-size
The `TArray` constructor in `SemType` supports "fully-dynamic" arrays and
"fully-static" ones, but not assumed-size arrays, which are a holdover
from older Fortran standards where only the final dimension is dynamic.
We replace the `Dimension` type synonym with a list-like data type in
order to support them. Laziness (and thus performance) should be the same.